+++ be/market/src/main/java/com/carrot/market/auth/infrastructure/kakao/config/KakaoOauthConfig.java
@@ -1,0 +1,13 @@
+package com.carrot.market.auth.infrastructure.kakao.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "oauth.kakao")
+public record KakaoOauthConfig(
+	String redirectUri,
+	String clientId,
+	String clientSecret,
+	String[] scope
+) {
+}
+
