+++ be/market/src/main/java/com/carrot/market/auth/infrastructure/kakao/dto/KakaoToken.java
@@ -1,0 +1,16 @@
+package com.carrot.market.auth.infrastructure.kakao.dto;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record KakaoToken(
+	String tokenType,
+	String accessToken,
+	String idToken,
+	Integer expiresIn,
+	String refreshToken,
+	Integer refreshTokenExpiresIn,
+	String scope
+) {
+}
