+++ be/market/src/main/java/com/carrot/market/auth/presentation/dto/request/LoginRequest.java
@@ -1,0 +1,6 @@
+package com.carrot.market.auth.presentation.dto.request;
+
+public record LoginRequest(
+	String code
+) {
+}
