+++ be/market/src/main/java/com/carrot/market/global/filter/JwtAuthorizationFilter.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.util.PatternMatchUtils;
+import org.springframework.util.StringUtils;
 import org.springframework.web.cors.CorsUtils;
 
 import com.carrot.market.global.exception.domain.JwtException;
@@ -103,7 +104,11 @@ public class JwtAuthorizationFilter implements Filter {
 	}
 
 	private String extractToken(String authorization) {
-		return authorization.substring(BEARER_PREFIX_LENGTH).replace("\"", "");
+		if (StringUtils.hasText(authorization) && authorization.startsWith(TOKEN_PREFIX)) {
+			return authorization.substring(BEARER_PREFIX_LENGTH).replace("\"", "");
+		}
+
+		throw new IllegalArgumentException();
 	}
 
 	private void sendErrorApiResponse(ServletResponse response, RuntimeException ex) throws IOException {
