+++ be/market/src/main/java/com/carrot/market/member/infrastructure/WishListRepository.java
@@ -1,15 +1,22 @@
 package com.carrot.market.member.infrastructure;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import com.carrot.market.member.domain.Member;
 import com.carrot.market.member.domain.WishList;
 
 @Repository
 public interface WishListRepository extends JpaRepository<WishList, Long> {
-	// @Query("select count(wl.id) > 0 from WishList as wl where wl.member.id = :member_id and wl.product.id = :product_id")
 	Boolean existsWishListByMemberIdAndProductId(@Param("member_id") Long memberId,
 		@Param("product_id") Long productId);
 
+	@Query("select wl from WishList wl "
+		+ "join fetch wl.category c "
+		+ "join fetch wl.product p")
+	List<WishList> findWishListByMember(Member member);
 }
