--- be/market/src/main/java/com/carrot/market/oauth/domain/OauthMember.java
@@ -1,18 +1,0 @@
-package com.carrot.market.oauth.domain;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class OauthMember {
-	@JsonProperty("socialId")
-	private String socialId;
-	@JsonProperty("imageUrl")
-	private String imageUrl;
-
-}
