--- be/market/src/main/java/com/carrot/market/oauth/presentation/annotation/OauthLogin.java
@@ -1,11 +1,0 @@
-package com.carrot.market.oauth.presentation.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface OauthLogin {
-}
