--- be/market/src/main/java/com/carrot/market/oauth/presentation/dto/request/LoginRequest.java
@@ -1,6 +1,0 @@
-package com.carrot.market.oauth.presentation.dto.request;
-
-public record LoginRequest(
-	String code
-) {
-}
