+++ be/market/src/main/java/com/carrot/market/product/application/ProductService.java
@@ -1,12 +1,17 @@
 package com.carrot.market.product.application;
 
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.carrot.market.member.domain.Member;
+import com.carrot.market.member.domain.WishList;
+import com.carrot.market.member.infrastructure.MemberRepository;
 import com.carrot.market.member.infrastructure.WishListRepository;
 import com.carrot.market.product.application.dto.response.CategoryDto;
 import com.carrot.market.product.application.dto.response.DetailPageServiceDto;
@@ -34,6 +39,7 @@ public class ProductService {
 	private final ProductRepository productRepository;
 	private final ProductImageRepository productImageRepository;
 	private final WishListRepository wishListRepository;
+	private final MemberRepository memberRepository;
 
 	public DetailPageServiceDto getMainPage(Long locationId, Long categoryId, Long next, int size) {
 
@@ -118,20 +124,26 @@ public class ProductService {
 	}
 
 	public WishListDetailDto getWishList(Long categoryId, Long memberId, Long next, int size) {
-		List<Category> categoryByMemberId = productRepository.findCategoryByMemberId(memberId);
+		Member member = memberRepository.findById(memberId).get();
+		Set<Category> categories = wishListRepository.findWishListByMember(member)
+			.stream()
+			.map(WishList::getCategory)
+			.collect(Collectors.toSet());
+
 		DetailPageSliceRequestDto build = DetailPageSliceRequestDto.builder()
 			.wishMemberId(memberId)
 			.categoryId(categoryId)
 			.nextProductId(next)
 			.pageSize(size)
 			.build();
-		Slice<DetailPageSliceResponseDto> byMyDetailPageSliceRequestDto = queryProductRepository.findByMyDetailPageSliceRequestDto(
-			build);
-		List<DetailPageSliceResponseDto> products = byMyDetailPageSliceRequestDto.getContent();
+
+		List<DetailPageSliceResponseDto> products = queryProductRepository.findByMyDetailPageSliceRequestDto(
+			build).getContent();
+
 		Long contentNextId = getContentNextId(products, size);
 		products = removeLastIfProductsSizeOverPageSize(products, size);
 
-		return WishListDetailDto.from(categoryByMemberId, products, contentNextId);
+		return WishListDetailDto.from(categories, products, contentNextId);
 	}
 
 }
