+++ be/market/src/main/java/com/carrot/market/product/application/dto/response/WishListDetailDto.java
@@ -1,6 +1,8 @@
 package com.carrot.market.product.application.dto.response;
 
+import java.util.Comparator;
 import java.util.List;
+import java.util.Set;
 
 import com.carrot.market.product.domain.Category;
 import com.carrot.market.product.infrastructure.dto.response.DetailPageSliceResponseDto;
@@ -11,8 +13,13 @@ public record WishListDetailDto(
 	Long nextId
 ) {
 
-	public static WishListDetailDto from(List<Category> categories, List<DetailPageSliceResponseDto> products,
+	public static WishListDetailDto from(Set<Category> categories, List<DetailPageSliceResponseDto> products,
 		Long nextId) {
-		return new WishListDetailDto(categories.stream().map(CategoryDto::from).toList(), products, nextId);
+		List<CategoryDto> categoryDtos = categories.stream()
+			.map(CategoryDto::from)
+			.sorted(Comparator.comparing(CategoryDto::id))
+			.toList();
+
+		return new WishListDetailDto(categoryDtos, products, nextId);
 	}
 }
