+++ be/market/src/main/java/com/carrot/market/product/infrastructure/ProductRepository.java
@@ -1,7 +1,5 @@
 package com.carrot.market.product.infrastructure;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -9,7 +7,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import com.carrot.market.product.application.dto.response.ProductSellerDetaillDto;
-import com.carrot.market.product.domain.Category;
 import com.carrot.market.product.domain.Product;
 
 @Repository
@@ -31,10 +28,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
 	Long findHitsById(@Param("id") Long id);
 
 	@Modifying(clearAutomatically = true)
-	@Query("update Product p set p.productDetails.hits = :hits where p.id = :id")
+	@Query("update Product p set p.productDetails.hits = p.productDetails.hits + :hits "
+		+ "where p.id = :id")
 	void applyHitsToRDB(@Param("id") Long id, @Param("hits") Long hits);
-
-	@Query("select p.category from Product p  join  p.category c join WishList as wl on wl.product = p where wl.member.id = :memberId")
-	List<Category> findCategoryByMemberId(@Param("memberId") Long memberId);
-
 }
