+++ be/market/src/main/java/com/carrot/market/product/infrastructure/dto/request/DetailPageSliceRequestDto.java
@@ -15,5 +15,4 @@ public record DetailPageSliceRequestDto(
 	@Builder
 	public DetailPageSliceRequestDto {
 	}
-
 }
